Removing sudoers support
This removes all tasks related to modifying the sudoers file during bootstrapping.

This role now requires the underlying OS to implement the required features this role used to ensure directly. I.e. a base image should include these features when it is built, so that Droplets are created with these features rather than adding them when it is provisioned.

This reduces the complexity of this role significantly (the lineinfile module tasks didn't work correctly anyway). Assuming users are using the BAS maintained base image they will notice no difference (other than needing to recreate Droplets).

As this is definitely a breaking change, this forms a new major release (1.0.0).